Refactor content provider architecture in Osirion CMS
This commit introduces a new `IContentProviderRegistry` interface for better management of content providers. The `ContentProviderFactory` and `ContentProviderManager` classes have been updated to utilize this registry, enhancing provider organization and default settings management.

The `ContentProviderBase` class now includes in-memory caching to improve performance, while the `MarkdownProcessor` class has been reintroduced with enhanced rendering and front matter extraction capabilities.

Several unused interfaces and classes have been removed, streamlining the codebase, and the `ServiceCollectionExtensions` class has been refactored for simpler service registration.

These changes aim to improve the modularity, maintainability, and performance of the Osirion CMS.
