+++ examples/bs/Osirion.Blazor.Example/Osirion.Blazor.Example/Components/Pages/CmsDemo.razor
@@ -12,7 +12,7 @@
         <div class="col-md-8">
             <div class="card shadow-sm">
                 <div class="card-body">
-                    <ContentViewer Path="@_contentPath" ProviderId="@_providerId" />
+                    <ContentView Path="@_contentPath"  />
                 </div>
             </div>
         </div>
