+++ examples/bs/Osirion.Blazor.Example/Osirion.Blazor.Example/appsettings.json
@@ -5,5 +5,66 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "Osirion": {
+    "Cms": {
+      "GitHub": {
+        "Owner": "obrana-boranija",
+        "Repository": "hexavera-blog",
+        "Branch": "main",
+        "ContentPath": "examples/content",
+        "ApiToken": "",
+        "ProviderId": "github-main",
+        "IsDefault": true,
+        "EnableCaching": true,
+        "CacheDurationMinutes": 30,
+        "EnableLocalization": true,
+        "DefaultLocale": "en",
+        "SupportedLocales": [ "en", "es", "fr" ],
+        "SupportedExtensions": [ ".md", ".markdown" ],
+        "ValidateContent": true,
+        "CommitterName": "CMS System",
+        "CommitterEmail": "cms@example.com"
+      },
+      //"FileSystem": {
+      //  "BasePath": "C:\\Content\\cms-files",
+      //  "WatchForChanges": true,
+      //  "PollingIntervalMs": 30000,
+      //  "IncludeSubdirectories": true,
+      //  "CreateDirectoriesIfNotExist": true,
+      //  "IncludePatterns": [ "**/*.md", "**/*.markdown" ],
+      //  "ExcludePatterns": [ "**/node_modules/**", "**/bin/**", "**/obj/**" ],
+      //  "ProviderId": "filesystem-local",
+      //  "IsDefault": false,
+      //  "EnableCaching": true,
+      //  "CacheDurationMinutes": 15,
+      //  "SupportedExtensions": [ ".md", ".markdown" ],
+      //  "EnableLocalization": false,
+      //  "DefaultLocale": "en",
+      //  "SupportedLocales": [ "en" ],
+      //  "ContentRoot": "/content"
+      //},
+      "Cache": {
+        "Enabled": true,
+        "StaleTimeMinutes": 5,
+        "MaxAgeMinutes": 60,
+        "UseStaleWhileRevalidate": true,
+        "SizeLimitMB": 64,
+        "CompactionPercentage": 0.25
+      }
+    }
+  },
+  "CmsAdmin": {
+    "Owner": "your-github-username",
+    "DefaultRepository": "your-content-repository",
+    "DefaultBranch": "main",
+    "AllowBranchCreation": true,
+    "AllowPullRequests": true,
+    "AllowFileDelete": true,
+    "AllowedRepositories": [],
+    "CommitterName": "CMS Admin",
+    "CommitterEmail": "admin@example.com",
+    "AllowedFileExtensions": [ ".md", ".markdown" ],
+    "ApiToken": "your-github-personal-access-token"
+  }
 }
