+++ src/Osirion.Blazor.Cms.Infrastructure/DependencyInjection/ApplicationBuilderExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Osirion.Blazor.Cms.Domain.Interfaces;
 
 namespace Osirion.Blazor.Cms.Infrastructure.DependencyInjection;
@@ -12,13 +13,22 @@ public static class ApplicationBuilderExtensions
     /// <summary>
     /// Initializes Osirion CMS providers
     /// </summary>
-    public static IApplicationBuilder InitializeOsirionCms(this IApplicationBuilder app)
+    public static IApplicationBuilder UseOsirionCms(this IApplicationBuilder app)
     {
         using var scope = app.ApplicationServices.CreateScope();
         var initializer = scope.ServiceProvider.GetRequiredService<IContentProviderInitializer>();
 
         // Initialize providers synchronously at startup
-        initializer.InitializeProvidersAsync().GetAwaiter().GetResult();
+        try
+        {
+            initializer.InitializeProvidersAsync().GetAwaiter().GetResult();
+        }
+        catch (Exception ex)
+        {
+            // Log the error but don't crash the application
+            var logger = scope.ServiceProvider.GetRequiredService<ILogger<IContentProviderInitializer>>();
+            logger.LogError(ex, "Error initializing content providers");
+        }
 
         return app;
     }
