+++ src/Osirion.Blazor.Cms/Extensions/ServiceCollectionExtensions.cs
@@ -1,19 +1,7 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Osirion.Blazor.Cms.Application.Commands;
-using Osirion.Blazor.Cms.Application.Queries;
-using Osirion.Blazor.Cms.Domain.Interfaces;
 using Osirion.Blazor.Cms.Domain.Options;
-using Osirion.Blazor.Cms.Domain.Repositories;
-using Osirion.Blazor.Cms.Domain.Services;
-using Osirion.Blazor.Cms.Infrastructure.Caching;
 using Osirion.Blazor.Cms.Infrastructure.DependencyInjection;
-using Osirion.Blazor.Cms.Infrastructure.Extensions;
-using Osirion.Blazor.Cms.Infrastructure.Factories;
-using Osirion.Blazor.Cms.Infrastructure.FileSystem;
-using Osirion.Blazor.Cms.Infrastructure.GitHub;
-using Osirion.Blazor.Cms.Infrastructure.Markdown;
-using Osirion.Blazor.Cms.Infrastructure.Services;
 
 namespace Osirion.Blazor.Cms;
 
@@ -32,38 +20,39 @@ public static class ServiceCollectionExtensions
         this IServiceCollection services,
         IConfiguration configuration)
     {
-        // Required services
-        services.AddMemoryCache();
-        services.AddScoped<IMarkdownProcessor, MarkdownProcessor>();
-        services.AddScoped<IMarkdownRendererService, MarkdownRendererService>();
-        services.AddScoped<IStateStorageService, LocalStorageService>();
 
-        // Configuration
-        services.Configure<GitHubOptions>(configuration.GetSection(GitHubOptions.Section));
-        services.Configure<FileSystemOptions>(configuration.GetSection(FileSystemOptions.Section));
-        services.Configure<CacheOptions>(configuration.GetSection(CacheOptions.Section));
+        services.AddCms(configuration);
+        //// Required services
+        //services.AddMemoryCache();
+        //services.AddScoped<IMarkdownProcessor, MarkdownProcessor>();
+        //services.AddScoped<IMarkdownRendererService, MarkdownRendererService>();
+        //services.AddScoped<IStateStorageService, LocalStorageService>();
 
-        // Register HTTP client
-        services.AddHttpClient<IGitHubApiClient, GitHubApiClient>();
+        //// Configuration
+        //services.Configure<GitHubOptions>(configuration.GetSection(GitHubOptions.Section));
+        //services.Configure<FileSystemOptions>(configuration.GetSection(FileSystemOptions.Section));
+        //services.Configure<CacheOptions>(configuration.GetSection(CacheOptions.Section));
 
-        // Register base repositories
-        services.AddScoped<GitHubContentRepository>();
-        services.AddScoped<GitHubDirectoryRepository>();
-        services.AddScoped<FileSystemContentRepository>();
-        services.AddScoped<FileSystemDirectoryRepository>();
+        //// Register HTTP client
+        //services.AddHttpClient<IGitHubApiClient, GitHubApiClient>();
 
-        // Register factory services
-        services.AddSingleton<CacheDecoratorFactory>();
-        services.AddScoped<IRepositoryFactory, RepositoryFactory>();
-        services.AddScoped<IUnitOfWorkFactory, UnitOfWorkFactory>();
+        //// Register base repositories
+        //services.AddScoped<GitHubContentRepository>();
+        //services.AddScoped<GitHubDirectoryRepository>();
+        //services.AddScoped<FileSystemContentRepository>();
+        //services.AddScoped<FileSystemDirectoryRepository>();
 
-        // Register CQRS services
-        services.AddScoped<ICommandDispatcher, CommandDispatcher>();
-        services.AddScoped<IQueryDispatcher, QueryDispatcher>();
+        //// Register factory services
+        //services.AddSingleton<CacheDecoratorFactory>();
+        //services.AddScoped<IRepositoryFactory, RepositoryFactory>();
+        //services.AddScoped<IUnitOfWorkFactory, UnitOfWorkFactory>();
 
-        // Provider services
-        services.AddScoped<IContentProviderFactory, ContentProviderFactory>();
-        services.AddScoped<IContentProviderManager, ContentProviderManager>();
+        //// Register CQRS services
+        //services.AddScoped<ICommandDispatcher, CommandDispatcher>();
+        //services.AddScoped<IQueryDispatcher, QueryDispatcher>();
+
+        //// Provider services
+        //services.AddScoped<IContentProviderManager, ContentProviderManager>();
 
         return services;
     }
@@ -71,15 +60,16 @@ public static class ServiceCollectionExtensions
     /// <summary>
     /// Adds configuration for GitHub content provider
     /// </summary>
-    public static IServiceCollection AddGitHubContentProvider(
+    public static IServiceCollection AddOsirionGitHubContentProvider(
         this IServiceCollection services,
         Action<GitHubOptions>? configure = null)
     {
-        // Allow additional configuration
-        if (configure != null)
-        {
-            services.Configure(configure);
-        }
+        //services.AddGitHubContentProvider(configure);
+        //// Allow additional configuration
+        //if (configure != null)
+        //{
+        //    services.Configure(configure);
+        //}
 
         return services;
     }
@@ -87,10 +77,11 @@ public static class ServiceCollectionExtensions
     /// <summary>
     /// Adds configuration for file system content provider
     /// </summary>
-    public static IServiceCollection AddFileSystemContentProvider(
+    public static IServiceCollection AddOsirionFileSystemContentProvider(
         this IServiceCollection services,
         Action<FileSystemOptions>? configure = null)
     {
+        //services.AddFileSystemContentProvider(configure);
         // Allow additional configuration
         if (configure != null)
         {
