+++ src/Osirion.Blazor.Cms/Osirion.Blazor.Cms.csproj
@@ -43,6 +43,7 @@
 	<ItemGroup>
 	  <ProjectReference Include="..\Osirion.Blazor.Cms.Admin\Osirion.Blazor.Cms.Admin.csproj" />
 		<ProjectReference Include="..\Osirion.Blazor.Cms.Front\Osirion.Blazor.Cms.Front.csproj" />
+		<ProjectReference Include="..\Osirion.Blazor.Cms.Infrastructure\Osirion.Blazor.Cms.Infrastructure.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
